+++ example/lib/core/data/data_sources/features/remote/features_remote_data_source_implementation.dart
@@ -1,0 +1,3 @@
+import 'features_remote_data_source.dart';
+
+class FeaturesRemoteDataSourceImplementation extends FeaturesRemoteDataSource {}