+++ lib/src/modules/graphql_mock_generator/operation_definition_node_visitor.dart
@@ -1,0 +1,12 @@
+import 'package:gql/ast.dart';
+
+class OperationDefinitionNodeVisitor extends RecursiveVisitor {
+  late final String name;
+  late final String operationTypeName;
+
+  @override
+  void visitOperationDefinitionNode(OperationDefinitionNode node) {
+    name = node.name?.value ?? '';
+    operationTypeName = node.type.name;
+  }
+}