+++ lib/src/modules/graphql_mock_generator/variable_definition_node_visitor.dart
@@ -1,0 +1,10 @@
+import 'package:gql/ast.dart';
+
+class VariableDefinitionNodeVisitor extends RecursiveVisitor {
+  late List<VariableDefinitionNode> variables;
+
+  @override
+  void visitVariableDefinitionNode(VariableDefinitionNode node) {
+    super.visitVariableDefinitionNode(node);
+  }
+}