+++ test/mocks/graphql_json_mocks.dart
@@ -1,0 +1,5 @@
+
+
+Map<String,dynamic> get featureTogglesMock => 
+  {"features":{"asset_withdraw_on_trading":true,"intercom":true}};
+